+++ CHANGELOG.md
@@ -49,6 +49,7 @@ end
 - [@kddeisz] - Don't convert into string literal arrays if the elements have brackets.
 - [@kddeisz] - Ensure you break the parent when there is an assignment in the predicate of a loop.
 - [@kddeisz] - Fix up a bug with keyword aliases, e.g., `alias in within`.
+- [@kddeisz] - Force using braces for regex if a regex starts with a blank space.
 
 # [0.22.0] - 2020-12-08